Remove seedVec from searchAllSeedsDoAll
Small refactor to SeedAlignerSearchData to support what SeedAndCache
(seedVec) used to support.

The char sequence pointer may be removable, but causes some issues at
the moment. Some commented out draft for it is there

aligner_cache.h QKey struct change to use 32->8bits for length. Same
output as before, and this is what CacheAndSeed use for length.

+++ aligner_cache.h
@@ -245,7 +245,7 @@ struct QKey {
 #endif
 
 	uint64_t seq; // sequence
-	uint32_t len; // length of sequence
+	uint8_t len; // length of sequence
 };
 
 class AlignmentCache;
